disable terminate button when the instance is not in a running state and refresh the screen on terminate that has been confirmed in a modal

+++ client/app/environments/dialogs/asg/asgInstances.js
@@ -9,7 +9,7 @@ angular.module('EnvironmentManager.environments').component('asgInstances', {
     asgState: '<'
   },
   controllerAs: 'vm',
-  controller: function ($uibModal, $http) {
+  controller: function ($uibModal, $http, $scope, modal) {
     var vm = this;
     vm.dataLoading = false;
 
@@ -26,11 +26,29 @@ angular.module('EnvironmentManager.environments').component('asgInstances', {
     };
 
     vm.terminateInstance = function (instanceId, accountName) {
-      //TODO: What to we want to do in the UI with this operation?
-      $http.delete('/api/v1/instances/' + instanceId + '?account=' + accountName)
-        .then(function (response) {
-          alert(JSON.stringify(response));
-        });
+      var modalParameters = {
+        title: 'Terminate instance?',
+        message: 'Are you sure you want to terminate <strong>' + instanceId + '</strong>?',
+        action: 'Terminate',
+        severity: 'Danger',
+        acknowledge: 'I am sure I want to terminate this instance: ' + instanceId + '.'
+      };
+
+      return modal.confirmation(modalParameters).then(function () {
+        return $http.delete('/api/v1/instances/' + instanceId + '?account=' + accountName)
+          .then(() => {
+            $scope.$emit('InstanceTerminating');
+          });
+      });
+    };
+
+    vm.notReadyForTerminate = function (instance) {
+      if (instance.LifecycleState.toLowerCase() !== 'inservice'
+        || instance.State.toLowerCase() !== 'running') {
+        return true;
+      }
+
+      return false;
     };
   }
 });
